requirements.txt
loguru==0.7.2
numpy==1.24.4
pandas==2.0.3
scikit_learn==1.3.2
scipy==1.10.1
tqdm==4.66.2
opencv-python==4.9.0.80
pyyaml==6.0.1
mkl==2024.0.0
matplotlib==3.7.5